more splitting tests in separate files
